fix(settings): split shortcut helpers for line check
